requirements.txt
streamlit
openai
python-docx
PyPDF2
scikit-learn
pandas
numpy
